Refactor flow_test.go
